# Optimize memoization