--- a/.changeset/silver-impalas-happen.md
+++ b/.changeset/silver-impalas-happen.md
@@ -1,0 +1,5 @@
+---
+"inngest": patch
+---
+
+Optimize memoization of steps, providing a performance improvement of up 48x for very high step counts